Add files via upload
seperated the tightly coupled product object and builder interface. it was a hierarchy mistake and now doing best to memorize proper hierarchy and process flow to avoid another tightly coupled implememtnation. previous worked but was not very elegant in implementation. modified director to do more with the builder as it was tedious to have the client make so many calls to builder and director.